11.09.2025 Обновление работы с переменными окружения

+++ embeddings_service/schemas.py
@@ -25,7 +25,7 @@ class HealthCheck(BaseModel):
     status: Literal["healthy", "failed"] = "healthy"
     model: str
     device: str = "cpu"
-    uptime: float | None = None
+    uptime: int | None = None
 
 
 class EmbeddingRequest(BaseModel):
